Downgraded SWserial and fixed bug with TX HW serial
Since SWserial included in core 2.4.2 and newer was causing lots of issues (data corruption), the old version patched for ESPeasy was added again.
For example, the Eastron SDM120 plugin was experiencing 20 - 30% of the data with CRC errors. Now it is near zero.
Also transparent selection of HWserial/SWserial is now working like it should.