[2.0.7] Make work-around for ESP32 not detaching pins between end/begin
ESP32 does have some issue where it is not properly detaching the RX/TX pins when calling `end()`
This happens especially when the RX pin receives data when calling `end()`.

Problem is that the serial port will no longer work after calling `end()` followed by `begin()` until it reboots.

This very messy work-around is to cache the used pins and not calling `begin()` on the port if using the same pins for a specific port.

Should be fixed with: https://github.com/espressif/arduino-esp32/pull/5385
But that core version cannot yet be used due to other issues, like a bug with web authentication and issues with WiFi and ESP-NOW.